@@ -7,119 +7,126 @@
 
 import Foundation
 
-//public enum APIEndpoint { }
+struct AnyEncodable: Encodable {
+    private let value: Encodable
 
-/// Defines all data needed to build the URL Request with.
-//public struct APIEndpoint<T> {
-//
-//    /// The path to the endpoint.
-//    let path: String
-//
-//    /// The HTTP Method to use for the request.
-//    let method: HTTPMethod
-//
-//    /// The parameters to send with the request. Can be `nil`.
-//    let parameters: [String: Any]?
-//
-//    /// The body to send with the request. Can be `nil`.
-//    let body: Data?
-//
-//    init(path: String, method: HTTPMethod = .get, parameters: [String: Any]? = nil, body: Data? = nil) {
-//        self.path = path
-//        self.method = method
-//        self.parameters = parameters
-//        self.body = body
-//    }
-//}
+    init(_ value: Encodable) {
+        self.value = value
+    }
+
+    func encode(to encoder: Encoder) throws {
+        try value.encode(to: encoder)
+    }
+}
+
+/// Credits to the https://github.com/kean/Get repository for this class.
+/// We've copied this over since it works nicely together with the CreateAPI OpenAPI generator.
+public struct Request<Response> {
+    public var method: String
+    public var path: String
+    public var query: [(String, String?)]?
+    var body: AnyEncodable?
+    public var headers: [String: String]?
+    public var id: String?
+
+    public init(method: String, path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) {
+        self.method = method
+        self.path = path
+        self.query = query
+        self.headers = headers
+    }
+
+    public init<U: Encodable>(method: String, path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) {
+        self.method = method
+        self.path = path
+        self.query = query
+        self.body = body.map(AnyEncodable.init)
+        self.headers = headers
+    }
+
+    public static func get(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "GET", path: path, query: query, headers: headers)
+    }
+
+    public static func post(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "POST", path: path, query: query, headers: headers)
+    }
+
+    public static func post<U: Encodable>(_ path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) -> Request {
+        Request(method: "POST", path: path, query: query, body: body, headers: headers)
+    }
+
+    public static func put(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "PUT", path: path, query: query, headers: headers)
+    }
+
+    public static func put<U: Encodable>(_ path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) -> Request {
+        Request(method: "PUT", path: path, query: query, body: body, headers: headers)
+    }
+
+    public static func patch(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "PATCH", path: path, query: query, headers: headers)
+    }
+
+    public static func patch<U: Encodable>(_ path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) -> Request {
+        Request(method: "PATCH", path: path, query: query, body: body, headers: headers)
+    }
+
+    public static func delete(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "DELETE", path: path, query: query, headers: headers)
+    }
+
+    public static func delete<U: Encodable>(_ path: String, query: [(String, String?)]? = nil, body: U?, headers: [String: String]? = nil) -> Request {
+        Request(method: "DELETE", path: path, query: query, body: body, headers: headers)
+    }
+
+    public static func options(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "OPTIONS", path: path, query: query, headers: headers)
+    }
+
+    public static func head(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "HEAD", path: path, query: query, headers: headers)
+    }
+
+    public static func trace(_ path: String, query: [(String, String?)]? = nil, headers: [String: String]? = nil) -> Request {
+        Request(method: "TRACE", path: path, query: query, headers: headers)
+    }
+}
 
 // MARK: - URLRequestConvertible
 
-//extension APIEndpoint {
-//
-//    /// Generates an URL based on the current endpoint in combination with the current API version.
-//    internal var url: URL {
-//        // swiftlint:disable:next force_unwrapping
-//        return URL(string: "https://api.appstoreconnect.apple.com/v1/")!.appendingPathComponent(path)
-//    }
-//
-//    /// Generates a request based on the current endpoint.
-//    public func asURLRequest() throws -> URLRequest {
-//        var urlRequest = URLRequest(url: url)
-//        urlRequest.httpMethod = method.rawValue
-//        urlRequest.encodeParameters(parameters)
-//
-//        if let body = body {
-//            if urlRequest.value(forHTTPHeaderField: "Content-Type") == nil {
-//                urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
-//            }
-//            urlRequest.httpBody = body
-//        }
-//        return urlRequest
-//    }
-//}
-//
-//// MARK: - Private
-//
-//private extension URLRequest {
-//
-//    mutating func encodeParameters(_ parameters: [String: Any]?) {
-//        guard
-//            let parameters = parameters, parameters.isEmpty == false,
-//            let url = url else { return }
-//
-//        func encode(_ value: Any) -> String {
-//            func percentEncode(_ string: String) -> String {
-//                return string.addingPercentEncoding(withAllowedCharacters: .afURLQueryAllowed) ?? string
-//            }
-//            switch value {
-//            case let intValue as Int:
-//                return percentEncode(String(intValue))
-//            case let stringValue as String:
-//                return percentEncode(stringValue)
-//            default:
-//                fatalError("Could not encode \(value)")
-//            }
-//        }
-//
-//        func query(_ parameters: [String: Any]) -> String {
-//            return parameters.sorted { $0.key < $1.key }.map { "\(encode($0))=\(encode($1))" }.joined(separator: "&")
-//        }
-//
-//        if httpMethod == HTTPMethod.get.rawValue {
-//            let newQueryToAppend = query(parameters)
-//            var urlComponents = URLComponents(url: url, resolvingAgainstBaseURL: false)
-//            let existingQuery = urlComponents?.percentEncodedQuery.map { $0 + "&" } ?? ""
-//            urlComponents?.percentEncodedQuery = existingQuery + newQueryToAppend
-//            self.url = urlComponents?.url
-//        } else {
-//            if value(forHTTPHeaderField: "Content-Type") == nil {
-//                setValue("application/x-www-form-urlencoded; charset=utf-8", forHTTPHeaderField: "Content-Type")
-//            }
-//
-//            httpBody = query(parameters).data(using: .utf8, allowLossyConversion: false)
-//        }
-//
-//    }
-//}
+extension Request {
 
-/// URLQueryAllowed CharacterSet extracted from Alamofire
-/// https://github.com/Alamofire/Alamofire
-fileprivate extension CharacterSet {
-    /// Creates a CharacterSet from RFC 3986 allowed characters.
-    ///
-    /// RFC 3986 states that the following characters are "reserved" characters.
-    ///
-    /// - General Delimiters: ":", "#", "[", "]", "@", "?", "/"
-    /// - Sub-Delimiters: "!", "$", "&", "'", "(", ")", "*", "+", ",", ";", "="
-    ///
-    /// In RFC 3986 - Section 3.4, it states that the "?" and "/" characters should not be escaped to allow
-    /// query strings to include a URL. Therefore, all "reserved" characters with the exception of "?" and "/"
-    /// should be percent-escaped in the query string.
-    static let afURLQueryAllowed: CharacterSet = {
-        let generalDelimitersToEncode = ":#[]@" // does not include "?" or "/" due to RFC 3986 - Section 3.4
-        let subDelimitersToEncode = "!$&'()*+,;="
-        let encodableDelimiters = CharacterSet(charactersIn: "\(generalDelimitersToEncode)\(subDelimitersToEncode)")
+    internal var baseURL: URL {
+        // swiftlint:disable:next force_unwrapping
+        return URL(string: "https://api.appstoreconnect.apple.com")!
+    }
 
-        return CharacterSet.urlQueryAllowed.subtracting(encodableDelimiters)
-    }()
+    private func makeURL(path: String, query: [(String, String?)]?) throws -> URL {
+        let url = baseURL.appendingPathComponent(path)
+        guard var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else {
+            throw URLError(.badURL)
+        }
+        if let query = query, !query.isEmpty {
+            components.queryItems = query.map(URLQueryItem.init)
+        }
+        guard let url = components.url else {
+            throw URLError(.badURL)
+        }
+        return url
+    }
+
+    /// Generates a request based on the current endpoint.
+    public func asURLRequest(encoder: JSONEncoder) throws -> URLRequest {
+        let url = try makeURL(path: path, query: query)
+        var urlRequest = URLRequest(url: url)
+        urlRequest.httpMethod = method
+
+        if let body = body {
+            urlRequest.httpBody = try encoder.encode(body)
+            urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
+        }
+
+        return urlRequest
+    }
 }
