@@ -11,10 +11,12 @@ import Foundation
 public struct Response<T> {
     public typealias StatusCode = Int
 
+    public let requestURL: URL?
     public let statusCode: Int
     public let data: T?
 
-    public init(statusCode: StatusCode, data: T?) {
+    public init(requestURL: URL?, statusCode: StatusCode, data: T?) {
+        self.requestURL = requestURL
         self.statusCode = statusCode
         self.data = data
     }
@@ -25,9 +27,6 @@ public protocol RequestExecutor {
 
     /// Performs a URLRequest and returns a result
     func execute(_ urlRequest: URLRequest, completion: @escaping (Result<Response<Data>, Swift.Error>) -> Void)
-
-    /// Retrieves the content of a given URL
-    func retrieve(_ url: URL, completion: @escaping (Result<Response<Data>, Swift.Error>) -> Void)
 
     /// Download report as file a given URL
     func download(_ urlRequest: URLRequest, completion: @escaping (Result<Response<URL>, Swift.Error>) -> Void)
